完成示例

+++ user/model/user.go
@@ -1,7 +1,7 @@
 package model
 
 type User struct {
-	Id        int    `json:"id"          sql:"id"`
+	Id        int64  `json:"id"          sql:"id"`
 	Username  string `json:"username"    sql:"username"`
 	LastName  string `json:"last_name"   sql:"last_name"`
 	FirstName string `json:"first_name"  sql:"first_name"`
